Factoring out configs and definitions
Suggestion to move the user config out to its own file along with clarifications where possible as to what the configurations mean- further cleanup to come - such as moving the main file into non-test and splitting the functionalities such that it will be easier to work with later. As a note acceleration scaling cfg val (set to 80) is not intuitive at all - I couldn't really figure out what it meant even glancing at where it's used.